Shuffle some files around
