+++ packages/thumbprint-react/utils/can-use-dom.js
@@ -1,0 +1,8 @@
+// https://stackoverflow.com/a/32598826/316602
+const canUseDOM = !!(
+    typeof window !== 'undefined' &&
+    window.document &&
+    window.document.createElement
+);
+
+export default canUseDOM;
